+++ extras/blurPrepare.py
@@ -5,6 +5,8 @@ import time
 from PIL import Image as Im
 from PIL import ImageFilter
 
+from extras import extrasUtil
+
 # Helper Variables
 slash = "\\" if os.name == 'nt' else "/"
 gauss_count = 0
@@ -28,13 +30,6 @@ def get_random_radius():
     return radius
 
 
-def check_file_count(in_folder):
-    file_count = 0
-    for root, dirs, files in os.walk(in_folder):
-        file_count += len(files)
-    return file_count
-
-
 def get_random_blur_type():
     global gauss_count, box_count
     random.seed(time.time_ns())
@@ -48,7 +43,7 @@ def get_random_blur_type():
 
 
 def process(input_folder):
-    file_count = check_file_count(input_folder)
+    file_count = extrasUtil.check_file_count(input_folder)
     index = 1
     rgb_index = 0
     failed_files = 0
