+++ extras/chromaPrepare.py
@@ -1,10 +1,8 @@
+import os
+
 from PIL import Image as Im
 from PIL import ImageCms as ImCms
 from PIL import ImageFilter
-import os
-import random
-import time
-import numpy as np
 
 slash = "\\" if os.name == 'nt' else "/"
