+++ extras/lumaPrepare.py
@@ -6,6 +6,8 @@ from PIL import Image as Im
 from PIL import ImageCms as ImCms
 from PIL import ImageFilter
 
+from extras import extrasUtil
+
 # Helper Variables
 
 slash = "\\" if os.name == 'nt' else "/"
@@ -36,15 +38,8 @@ def get_random_radius(a, b):
     return radius
 
 
-def check_file_count(in_folder):
-    file_count = 0
-    for root, dirs, files in os.walk(in_folder):
-        file_count += len(files)
-    return file_count
-
-
 def process(input_folder):
-    file_count = check_file_count(input_folder)
+    file_count = extrasUtil.check_file_count(input_folder)
     index = 1
     failed_files = 0
     skipped_files = 0
