+++ extras/monochromePrepare.py
@@ -3,21 +3,16 @@ import os
 from PIL import Image as Im
 from PIL import ImageOps as ImOps
 
+from extras import extrasUtil
+
 # Helper variables
 
 slash = "\\" if os.name == 'nt' else "/"
 valid_extensions = [".jpg", ".png", ".dds", ".bmp"]
 
 
-def check_file_count(in_folder):
-    file_count = 0
-    for root, dirs, files in os.walk(in_folder):
-        file_count += len(files)
-    return file_count
-
-
 def process(input_folder):
-    file_count = check_file_count(input_folder)
+    file_count = extrasUtil.check_file_count(input_folder)
     index = 1
     failed_files = 0
     skipped_files = 0
