+++ prepare_dataset_1x.py
@@ -1,12 +1,13 @@
 import random
 import time
-from os import walk, path, makedirs, listdir, name
+from os import path, makedirs, listdir, name
 from shutil import copyfile
 
 from PIL import Image as Im
 from PIL import ImageFile
 
 import select_tiles
+from extras import extrasUtil
 
 # Helper Variables and Flags
 
@@ -38,13 +39,6 @@ def get_random_number(start, end):
     # Use time as a seed, makes it more randomized
     random.seed(time.time_ns())
     return random.randint(start, end)
-
-
-def check_file_count(in_folder):
-    file_count = 0
-    for root, dirs, files in walk(in_folder):
-        file_count += len(files)
-    return file_count
 
 
 def process_image(image, filename):
@@ -95,7 +89,7 @@ def save():
 def main():
     print("Splitting dataset pictures...")
     rgb_index = 0
-    file_count = check_file_count(input_folder)
+    file_count = extrasUtil.check_file_count(input_folder)
     index = 1
     for filename in listdir(input_folder):
         time_var = int(time.time())
